fix : runningRecord 타입 수정

+++ backend/dduishu/src/main/java/com/b208/dduishu/domain/character/entity/CharacterLevel.java
@@ -26,7 +26,7 @@ public class CharacterLevel {
         this.exp = exp;
     }
 
-    public void addExp(int exp ) {
+    public void addExp(float exp ) {
         this.exp += exp;
         checkLevel();
     }
